+++ starsky/starskytest/starsky.feature.search/ViewModels/SearchViewModelTest.cs
@@ -5,7 +5,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.feature.search.ViewModels;
 using starsky.foundation.database.Models;
 
-namespace starskytest.starsky.foundation.search.ViewModels;
+namespace starskytest.starsky.feature.search.ViewModels;
 
 [TestClass]
 public sealed class SearchViewModelTest
@@ -14,14 +14,14 @@ public sealed class SearchViewModelTest
 	public void SearchViewModelTest_TestIfTrash()
 	{
 		var test = new SearchViewModel { SearchQuery = TrashKeyword.TrashKeywordString };
-		Assert.AreEqual(PageViewType.PageType.Trash.ToString(), test.PageType);
+		Assert.AreEqual(nameof(PageViewType.PageType.Trash), test.PageType);
 	}
 
 	[TestMethod]
 	public void SearchViewModelTest_Default()
 	{
 		var test = new SearchViewModel();
-		Assert.AreEqual(PageViewType.PageType.Search.ToString(), test.PageType);
+		Assert.AreEqual(nameof(PageViewType.PageType.Search), test.PageType);
 	}
 
 	[TestMethod]